apps/lib/http_server.{c,h}: clean up logging and move it to log.{c,h}
